code review comment updates
